+++ pkg/mcp/server.go
@@ -160,7 +160,6 @@ var (
 	allTools = []string{
 		ToolGet,
 		ToolList,
-		ToolChildren,
 		ToolSearch,
 		ToolTargets,
 		ToolID,
@@ -182,7 +181,6 @@ var (
 	readTools = []string{
 		ToolGet,
 		ToolList,
-		ToolChildren,
 		ToolSearch,
 		ToolTargets,
 		ToolID,
@@ -213,7 +211,6 @@ var (
 	aliasMap = map[string]string{
 		"get":             ToolGet,
 		"list":            ToolList,
-		"children":        ToolChildren,
 		"search":          ToolSearch,
 		"targets":         ToolTargets,
 		"id":              ToolID,